feat: big number ux ui changes
