+++ packages/frontend/src/components/VisualizationConfigs/BigNumberConfig/ConfigTabs/index.tsx
@@ -1,0 +1,29 @@
+import { MantineProvider, Tabs } from '@mantine/core';
+import { memo } from 'react';
+import { themeOverride } from '../../theme';
+import { Comparison } from '../Comparison';
+import { Layout } from '../Layout';
+
+export const ConfigTabs = memo(() => {
+    return (
+        <MantineProvider inherit theme={themeOverride}>
+            <Tabs defaultValue="layout" keepMounted={false}>
+                <Tabs.List mb="sm">
+                    <Tabs.Tab px="sm" value="layout">
+                        Layout
+                    </Tabs.Tab>
+                    <Tabs.Tab px="sm" value="comparison">
+                        Comparison
+                    </Tabs.Tab>
+                </Tabs.List>
+
+                <Tabs.Panel value="layout">
+                    <Layout />
+                </Tabs.Panel>
+                <Tabs.Panel value="comparison">
+                    <Comparison />
+                </Tabs.Panel>
+            </Tabs>
+        </MantineProvider>
+    );
+});
