+++ packages/frontend/src/components/VisualizationConfigs/BigNumberConfig/common/index.ts
@@ -1,0 +1,9 @@
+import { CompactConfigMap } from '@lightdash/common';
+
+export const StyleOptions = [
+    { value: '', label: 'None' },
+    ...Object.values(CompactConfigMap).map(({ compact, label }) => ({
+        value: compact,
+        label,
+    })),
+];
